Applied change originally by DWD:
This seems to add more locale categories to the header file. This
seems pointless because it's not required to build uclibc and
currently we don't build applications against uclibc.

Also note that originally DWD applied a similar change to
libc/sysdeps/linux/common/bits/locale.h

that was especially pointless! As the file is not even read as
far as I can tell!