Added a change originally by DWD.
This makes calls to open() and fcntl() made by _stdio_fopen()
go to an undefined externals __libc_open() and __libc_fcntl().

These are treated in specially in KLEE in
linkWithUclibc(llvm::Module *mainModule)

so that they get renamed to calls to KLEE's open() and fcntl()
models.

I don't know why KLEE treats these specially (maybe it's because
of the variable number of arguments issue?)